requirements.txt
ipykernel
jupyter
python-dotenv

# Building systems with chatGPT API
groq 
openai